+++ CHANGELOG.md
@@ -7,13 +7,18 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## [Unreleased] - YYYY-MM-DD
 #### Added
 #### Changed
-- Locking `react` to latest version (finding path to `react` folder hierarchically).
-- Saving `stats.json` object with build output.
-
 #### Deprecated
 #### Removed
 #### Fixed
 #### Security
+
+
+
+
+## [3.10.0] - 2016-06-18
+#### Changed
+- Locking `react` to latest version (finding path to `react` folder hierarchically).
+- Saving `stats.json` object with build output.